# Optimize unchecked indexing into chunks and chunks_mut