@the8472 2021-07-02
Grepping through the compiler source doesn't show any uses of `chunks` or `chunks_mut`, even less so in combination with `Zip`, so I don't really expect a change in performance, but maybe some dependencies do.

@bors try @rust-timer queue