@klensy 2021-07-02
Comparing CI build nightly with manually build stage1 can show unrelated info.