@the8472 2021-07-02
I'd expect the compiler internals to be built differently, but shouldn't the *output* of the compiler remain comparable?